more work on a type system
Now parses correctly the line:
struct lol FAR *LoL = &DATASTART;
so returned it into its unmodified form.
Unfortunately it forced me to add a few seemingly
innocent but useless typecasts to (int) in a few places.

+++ kernel/fcbfns.c
@@ -61,7 +61,7 @@ UBYTE FAR *FatGetDrvData(UBYTE drive, UBYTE * pspc, UWORD * bps, UWORD * nc)
   spc = DosGetFree(drive, NULL, bps, nc);
   if (spc != 0xffff)
   {
-    struct dpb FAR *dpbp = get_dpb(drive == 0 ? default_drive : drive - 1);
+    struct dpb FAR *dpbp = get_dpb(drive == 0 ? (int)default_drive : drive - 1);
     /* Point to the media desctriptor for this drive               */
     *pspc = (UBYTE)spc;
     if (dpbp == NULL)
